Moving dashboard stuff to its own package
