+++ src/main/java/de/zalando/ep/zalenium/proxy/TestingBotRemoteProxy.java
@@ -3,7 +3,7 @@ package de.zalando.ep.zalenium.proxy;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
-import de.zalando.ep.zalenium.util.TestInformation;
+import de.zalando.ep.zalenium.dashboard.TestInformation;
 import org.openqa.grid.common.RegistrationRequest;
 import org.openqa.grid.internal.Registry;
