+++ src/test/java/de/zalando/ep/zalenium/proxy/TestingBotRemoteProxyTest.java
@@ -3,6 +3,8 @@ package de.zalando.ep.zalenium.proxy;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
+import de.zalando.ep.zalenium.dashboard.Dashboard;
+import de.zalando.ep.zalenium.dashboard.TestInformation;
 import de.zalando.ep.zalenium.util.*;
 import org.hamcrest.CoreMatchers;
 import org.junit.*;
